refactor(apps/web): reorganize library albums component directory structure
