+++ apps/web/src/routes/app/library/artists/$id/add-content/album.tsx
@@ -1,4 +1,4 @@
-import { CreateAlbumForm } from '@/components/library-albums/CreateAlbumForm';
+import { CreateAlbumForm } from '@/components/library/albums/CreateAlbumForm';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { useCreateLibraryAlbum } from '@/hooks/api/library-albums/useCreateLibraryAlbum';
 import { useCreateAlbumForm } from '@/hooks/forms/useCreateAlbumForm';
